move redemption quote
